+++ compiler/src/compile/compile_tests.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use crate::test_utils::{assert_vals, exec, read_test};
+    use crate::test_utils::utils::{assert_vals, exec, read_test};
     use builtins::print::{dasm, prn};
     use compile_state::state::new_slosh_vm;
     use compile_state::state::SloshVmTrait;
